pyproject.toml
[project]
name = "dl_experiment_repo"
version = "0.1.0"
description = "Deep learning project using Hydra, Submitit, and W&B"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "torch",
    "torchvision",
    "hydra-core",
    "hydra-submitit-launcher",
    "wandb",
    "numpy",
    "pandas",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "jupyterlab"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_model"]
